+++ primitives/src/covenants/register.rs
@@ -58,21 +58,21 @@ impl RegisterCovenant {
 }
 
 impl Encodable for RegisterCovenant {
-    fn size(&self) -> usize {
-        let mut size = VarInt::from(4 as u64).encoded_size() as usize;
+    fn size(&self) -> u32 {
+        let mut size = VarInt::from(4 as u64).encoded_size();
         let name_hash_length = VarInt::from(32 as u64);
         let height_length = VarInt::from(4 as u64);
         let name_length = VarInt::from(self.record_data.len() as u64);
         let block_length = VarInt::from(32 as u64);
 
         //@todo fix.
-        size += name_hash_length.encoded_size() as usize;
-        size += height_length.encoded_size() as usize;
-        size += name_length.encoded_size() as usize;
-        size += block_length.encoded_size() as usize;
+        size += name_hash_length.encoded_size();
+        size += height_length.encoded_size();
+        size += name_length.encoded_size();
+        size += block_length.encoded_size();
         size += 32;
         size += 4;
-        size += self.record_data.len();
+        size += self.record_data.len() as u32;
         size += 32;
 
         size
