+++ primitives/src/covenants/revoke.rs
@@ -36,13 +36,13 @@ impl RevokeCovenant {
 }
 
 impl Encodable for RevokeCovenant {
-    fn size(&self) -> usize {
-        let mut size = VarInt::from(2 as u64).encoded_size() as usize;
+    fn size(&self) -> u32 {
+        let mut size = VarInt::from(2 as u64).encoded_size();
         let name_hash_length = VarInt::from(32 as u64);
         let height_length = VarInt::from(4 as u64);
 
-        size += name_hash_length.encoded_size() as usize;
-        size += height_length.encoded_size() as usize;
+        size += name_hash_length.encoded_size();
+        size += height_length.encoded_size();
         size += 32;
         size += 4;
