+++ primitives/src/covenants/transfer.rs
@@ -60,22 +60,22 @@ impl TransferCovenant {
 }
 
 impl Encodable for TransferCovenant {
-    fn size(&self) -> usize {
-        let mut size = VarInt::from(4 as u64).encoded_size() as usize;
+    fn size(&self) -> u32 {
+        let mut size = VarInt::from(4 as u64).encoded_size();
         //TODO because all these values are below 252
         let name_hash_length = VarInt::from(32 as u64);
         let height_length = VarInt::from(4 as u64);
         let version_length = VarInt::from(1 as u64);
         let address_length = VarInt::from(self.address.hash.as_hash().len() as u64);
 
-        size += name_hash_length.encoded_size() as usize;
-        size += height_length.encoded_size() as usize;
-        size += version_length.encoded_size() as usize;
-        size += address_length.encoded_size() as usize;
+        size += name_hash_length.encoded_size();
+        size += height_length.encoded_size();
+        size += version_length.encoded_size();
+        size += address_length.encoded_size();
         size += 32;
         size += 4;
         size += 1;
-        size += self.address.hash.as_hash().len();
+        size += self.address.hash.as_hash().len() as u32;
 
         size
     }
